move mailer to lib, add gmail config, closes #397
